requirements.txt
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
dj-database-url==2.1.0
psycopg2-binary==2.9.9
python-decouple==3.8
gunicorn==22.0.0
whitenoise==6.6.0
Pillow==10.3.0
